relocated the components
