+++ mean-stack/src/app/auth/signup/signup.component.ts
